+++ CHANGELOG.md
@@ -14,6 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   Added Jobs tab to agent page for viewing and managing background jobs
+-   Added ability to kill background jobs from the Jobs tab
+
 ## [3.3.0] - 2026-01-18
 
 ### Added